Fix found places where uio_resid is truncated to int.
Add the sysctl debug.iosize_max_clamp, enabled by default. Setting the
sysctl to zero allows to perform the SSIZE_MAX-sized i/o requests from
the usermode.

Discussed with:	bde, das (previous versions)
MFC after:	1 month


git-svn-id: svn+ssh://svn.freebsd.org/base/head@231949 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f

+++ sys/kern/kern_gzio.c
@@ -97,7 +97,7 @@ gzFile gz_open (path, mode, vp)
     gz_stream *s;
     char fmode[80]; /* copy of mode, without the compression level */
     char *m = fmode;
-    int resid;
+    ssize_t resid;
     int error;
     char buf[GZ_HEADER_LEN + 1];
 
@@ -342,7 +342,7 @@ local void putU32 (s, x)
 {
     uint32_t xx;
     off_t curoff = s->outoff;
-    int resid;
+    ssize_t resid;
 
 #if BYTE_ORDER == BIG_ENDIAN
     xx = bswap32(x);
